+++ IR/IR/DocumentTable.h
@@ -3,14 +3,14 @@
 
 class DocumentTable {
 private:
-	unsigned int documentNumber;
+	unsigned long long documentNumber;
 	bool finalized;
 	void * docTable; // it's a list at first, and a table in the end
 public:
 	DocumentTable();
 	DocumentTable(int documentNumber, DocumentMetaData * docTable);
 	unsigned int addDocument(DocumentMetaData documentMetaData);
-	unsigned int& getDocumentNumber();
+	unsigned long long& getDocumentNumber();
 	void finalize();
 	DocumentMetaData* getDocument(int documentIndex);
 	unsigned long long getMemorySize();
