+++ IR/IR/DocumentTerm.h
@@ -1,5 +1,5 @@
 #pragma once
 typedef struct DocumentTerm {
-	unsigned int documentIndex;
+	unsigned long long documentIndex;
 	unsigned int ftd=1; // frequency of term in the current document
 } DocumentTerm;
